refactor: extract getCollectionItems to separate utility file
