+++ packages/components/src/templates/next/layouts/Collection/utils/index.ts
@@ -3,3 +3,4 @@ export * from "./getFilteredItems"
 export * from "./getPaginatedItems"
 export * from "./updateAppliedFilters"
 export * from "./shouldShowDate"
+export * from "./getCollectionItems"
